requirements.txt
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.151.0
ollama==0.3.3
